+++ apps/web/src/app/play/checker/[id]/page.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { redirect } from "next/navigation";
 import * as z from "zod";
 
-import { monitorFlyRegionSchema } from "@openstatus/db/src/schema/shared/shared";
+import { monitorFlyRegionSchema } from "@openstatus/db/src/schema/shared";
 import { Separator } from "@openstatus/ui";
 
 import {
